refactor: domain/history を types/ + 関数別ファイルに整理
- 型定義は types/ サブディレクトリにコロケーション
  (HistoryItem, HistoryStore, ParsedSearchQuery, GroupingOptions, SearchOptions)
- 関数ファイルは「ファイル名 = 関数名」に統一
  (parseSearchQuery, filterByTerms, groupHistories, searchHistories)
- history-filter.ts は filterByTerms と groupHistories に分割
- SearchOptions が GroupingOptions と重複定義していたのを extends で解消
