+++ src/domain/history/filterByTerms.ts
@@ -1,11 +1,5 @@
-import { HistoryItem } from "./HistoryItem";
-import { ParsedSearchQuery } from "./SearchQuery";
-import { uniqBy } from "../../util/array";
-
-export interface GroupingOptions {
-  groupByUrl?: boolean;
-  groupByTitle?: boolean;
-}
+import { HistoryItem } from "./types/HistoryItem";
+import { ParsedSearchQuery } from "./types/ParsedSearchQuery";
 
 /**
  * 解析済みクエリのタームをすべて満たす履歴のみを残します (AND 検索)。
@@ -38,30 +32,6 @@ export function filterByTerms(
       return searchText.includes(parsedTerm.term);
     });
   });
-}
-
-/**
- * 同一 URL / 同一タイトルでグルーピングし、最新のものだけを残します。
- */
-export function groupHistories(
-  items: HistoryItem[],
-  options?: GroupingOptions,
-): HistoryItem[] {
-  if (!options?.groupByUrl && !options?.groupByTitle) {
-    return items;
-  }
-
-  const sorted = [...items].sort((a, b) => b.lastVisitTime - a.lastVisitTime);
-
-  const afterUrlGrouping = options.groupByUrl
-    ? uniqBy(sorted, (item) => item.url)
-    : sorted;
-
-  const afterTitleGrouping = options.groupByTitle
-    ? uniqBy(afterUrlGrouping, (item) => item.title || "")
-    : afterUrlGrouping;
-
-  return afterTitleGrouping;
 }
 
 if (import.meta.vitest) {
@@ -127,39 +97,6 @@ if (import.meta.vitest) {
       const items = [makeItem({ url: "not-a-url" })];
       const result = filterByTerms(items, [{ term: "anything", type: "site" }]);
       expect(result).toEqual([]);
-    });
-  });
-
-  describe("groupHistories", () => {
-    it("returns items as-is when no grouping requested", () => {
-      const items = [
-        makeItem({ url: "https://a.com", lastVisitTime: 1 }),
-        makeItem({ url: "https://a.com", lastVisitTime: 2 }),
-      ];
-      expect(groupHistories(items)).toEqual(items);
-      expect(groupHistories(items, {})).toEqual(items);
-    });
-
-    it("keeps the newest entry per URL when groupByUrl", () => {
-      const items = [
-        makeItem({ url: "https://a.com", title: "old", lastVisitTime: 1 }),
-        makeItem({ url: "https://a.com", title: "new", lastVisitTime: 2 }),
-        makeItem({ url: "https://b.com", title: "other", lastVisitTime: 1 }),
-      ];
-      const result = groupHistories(items, { groupByUrl: true });
-      expect(result).toHaveLength(2);
-      expect(result.find((i) => i.url === "https://a.com")?.title).toBe("new");
-    });
-
-    it("keeps the newest entry per title when groupByTitle", () => {
-      const items = [
-        makeItem({ url: "https://a.com", title: "Same", lastVisitTime: 1 }),
-        makeItem({ url: "https://b.com", title: "Same", lastVisitTime: 2 }),
-        makeItem({ url: "https://c.com", title: "Other", lastVisitTime: 1 }),
-      ];
-      const result = groupHistories(items, { groupByTitle: true });
-      expect(result).toHaveLength(2);
-      expect(result.find((i) => i.title === "Same")?.url).toBe("https://b.com");
     });
   });
 }
