+++ src/domain/history/types/GroupingOptions.ts
@@ -1,0 +1,4 @@
+export interface GroupingOptions {
+  groupByUrl?: boolean;
+  groupByTitle?: boolean;
+}
