+++ src/domain/history/types/ParsedSearchQuery.ts
@@ -1,0 +1,4 @@
+export type ParsedSearchQuery = {
+  term: string;
+  type: "text" | "site" | "exclude";
+};
